+++ src/main/java/Laws/PoissonLaw.java
@@ -15,11 +15,11 @@ public class PoissonLaw implements Law, Displayable {
         this.lambda = lambda;
     }
 
-    public double getProbabiliteDeX(int x_egal_i){
+    public double getProbabiliteDeX(double x_egal_i){
         double coef = Math.exp(-lambda);
         int denom = 0;
         try {
-            denom = (int)Calculation.factorielle(x_egal_i);
+            denom = (int)Calculation.factorielle((int)x_egal_i);
         } catch (Throwable e) {
             System.out.println(e.getMessage());
             System.out.printf("exiting -");
