+++ src/main/java/Laws/functions/Law.java
@@ -8,7 +8,7 @@ import Exceptions.CalculationException;
 public interface Law {
      String getName();
      String getParameters();
-     double getProbabiliteDeX(int x_egal_i) throws CalculationException;
+     double getProbabiliteDeX(double x_egal_i) throws CalculationException;
      double getEsperance();
      double getVariance();
 }
